[Fix] Should have checked if that last-minute change actually compiles...
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@22369 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/plugins/PlugInterface.cpp
@@ -382,7 +382,7 @@ bool IMixPlugin::MidiSend(uint32 midiCode)
 {
 	std::array<std::byte, 4> midiData;
 	memcpy(midiData.data(), &midiCode, 4);
-	return MidiSend(mpt::as_span(midiData.data(), std::min(midiData.size(), MIDIEvents::GetEventLength(mpt::byte_cast<uint8>(midiData[0])))));
+	return MidiSend(mpt::as_span(midiData.data(), std::min(static_cast<uint8>(midiData.size()), MIDIEvents::GetEventLength(mpt::byte_cast<uint8>(midiData[0])))));
 }
